+++ packages/server/auth/routes.js
@@ -495,8 +495,8 @@ router.post(
       res.redirect("/auth/login");
     } else if (result.error) {
       console.log({
-        result
-      })
+        result,
+      });
       req.flash("danger", result.error);
       const form = resetForm(req.body, req);
       form.errors = { password: result.error, confirm_password: result.error };
@@ -1334,7 +1334,8 @@ router.post(
       null,
       req.user,
       resultCollector,
-      req.user
+      req.user,
+      { req }
     );
     if (resultCollector.notify) {
       req.flash("warning", resultCollector.notify);
@@ -1452,7 +1453,7 @@ router.post(
  * @param {object} res
  * @returns {void}
  */
-const loginCallback = (req, res) => () => {
+const loginCallback = (req, res) => async () => {
   if (!req.user) return;
   if (!req.user.id) {
     res.redirect("/auth/signup_final_ext");
@@ -1460,7 +1461,28 @@ const loginCallback = (req, res) => () => {
   if (!req.user.email) {
     res.redirect("/auth/set-email");
   } else {
-    Trigger.emitEvent("Login", null, req.user);
+    const resultCollector = {};
+    await Trigger.runTableTriggers(
+      "Login",
+      null,
+      req.user,
+      resultCollector,
+      req.user,
+      { req }
+    );
+    if (resultCollector.notify) {
+      req.flash("warning", resultCollector.notify);
+    }
+    if (resultCollector.error) {
+      req.flash("error", resultCollector.error);
+    }
+    if (resultCollector.notify_success) {
+      req.flash("success", resultCollector.notify_success);
+    }
+    if (resultCollector.goto) {
+      res.redirect(resultCollector.goto);
+      return;
+    }
     req.flash("success", req.__("Welcome, %s!", req.user.email));
     if (req.cookies["login_dest"]) {
       res.clearCookie("login_dest");