+++ packages/server/serve.js
@@ -387,6 +387,7 @@ module.exports =
             Object.entries(cluster.workers).forEach(([wpid, w]) => {
               w.send(msg);
             });
+            workerDispatchMsg(msg); //also master
           };
 
           if (masterState.listeningTo.size < useNCpus)
@@ -431,6 +432,7 @@ module.exports =
           Object.entries(cluster.workers).forEach(([wpid, w]) => {
             w.send(msg);
           });
+          workerDispatchMsg(msg); //also master
         };
 
         cluster.on("exit", (worker, code, signal) => {
@@ -438,6 +440,9 @@ module.exports =
           addWorker(cluster.fork());
         });
       } else {
+        getState().sendMessageToWorkers = (msg) => {
+          workerDispatchMsg(msg); //also master
+        };
         await nonGreenlockWorkerSetup(appargs, port, host);
         runScheduler({
           port,
@@ -623,7 +628,7 @@ const setupSocket = (subdomainOffset, pruneSessionInterval, ...servers) => {
           callback({ status: "error", msg: err.message || "unknown error" });
         }
       };
-      if (tenant && tenant !== "public") db.runWithTenant(tenant, f);
+      if (tenant && tenant !== "public") await db.runWithTenant(tenant, f);
       else await f();
     });
 
@@ -647,7 +652,7 @@ const setupSocket = (subdomainOffset, pruneSessionInterval, ...servers) => {
             callback({ status: "error", msg: err.message || "unknown error" });
         }
       };
-      if (tenant && tenant !== "public") db.runWithTenant(tenant, f);
+      if (tenant && tenant !== "public") await db.runWithTenant(tenant, f);
       else await f();
     });
 
@@ -657,6 +662,8 @@ const setupSocket = (subdomainOffset, pruneSessionInterval, ...servers) => {
         get_tenant_from_req(socket.request, subdomainOffset) || "public";
       const f = async () => {
         try {
+          const enabled = getState().getConfig("enable_dynamic_updates", true);
+          if (!enabled) throw new Error("Dynamic updates are not enabled");
           const user = socket.request.user;
           if (!user) throw new Error("Not authorized");
           socket.join(`_${tenant}_dynamic_update_room`);
@@ -668,7 +675,7 @@ const setupSocket = (subdomainOffset, pruneSessionInterval, ...servers) => {
             callback({ status: "error", msg: err.message || "unknown error" });
         }
       };
-      if (tenant && tenant !== "public") db.runWithTenant(tenant, f);
+      if (tenant && tenant !== "public") await db.runWithTenant(tenant, f);
       else await f();
     });
 
@@ -676,12 +683,20 @@ const setupSocket = (subdomainOffset, pruneSessionInterval, ...servers) => {
       const tenant =
         get_tenant_from_req(socket.request, subdomainOffset) || "public";
       const f = async () => {
-        const socketIds = await getState().getConfig("joined_log_socket_ids");
-        const newSocketIds = socketIds.filter((id) => id !== socket.id);
-        await getState().setConfig("joined_log_socket_ids", newSocketIds);
+        const state = getState();
+        if (state) {
+          const socketIds = state.getConfig("joined_log_socket_ids");
+          const newSocketIds = socketIds.filter((id) => id !== socket.id);
+          await state.setConfig("joined_log_socket_ids", newSocketIds);
+        } else {
+          console.error("No state found in socket disconnect");
+          console.error("The current tenantSchema is:", db.getTenantSchema());
+          console.error("The current tenant from request is:", tenant);
+          console.error("node version:", process.version);
+        }
       };
-      if (tenant && tenant !== "public") db.runWithTenant(tenant, f);
-      else f();
+      if (tenant && tenant !== "public") await db.runWithTenant(tenant, f);
+      else await f();
     });
   });
 
@@ -729,8 +744,8 @@ const setupSocket = (subdomainOffset, pruneSessionInterval, ...servers) => {
             callback({ status: "error", msg: err.message || "unknown error" });
           }
         };
-        if (tenant && tenant !== "public") db.runWithTenant(tenant, f);
-        else f();
+        if (tenant && tenant !== "public") await db.runWithTenant(tenant, f);
+        else await f();
       }
     );
     socket.on("write_to_stream", async (data, callback) => {
@@ -792,8 +807,8 @@ const setupSocket = (subdomainOffset, pruneSessionInterval, ...servers) => {
             }
           });
       };
-      if (tenant && tenant !== "public") db.runWithTenant(tenant, f);
-      else f();
+      if (tenant && tenant !== "public") await db.runWithTenant(tenant, f);
+      else await f();
     });
   });
 };