+++ packages/server/wrapper.js
@@ -91,13 +91,13 @@ const get_headers = (req, version_tag, description, extras = []) => {
   const locale = req.getLocale();
   const stdHeaders = [
     {
-      headerTag: `<script>var _sc_loglevel = ${
+      headerTag: `<script>const _sc_loglevel = ${
         state.logLevel
       }, _sc_globalCsrf = "${req.csrfToken()}", _sc_version_tag = "${version_tag}"${
         locale ? `, _sc_locale = "${locale}"` : ""
       }, _sc_lightmode = ${JSON.stringify(
         state.getLightDarkMode?.(req.user) || "light"
-      )};</script>`,
+      )}, _sc_pageloadtag = Math.floor(Math.random() * 16777215).toString(16);</script>`,
     },
     { css: `/static_assets/${version_tag}/saltcorn.css` },
     { script: `/static_assets/${version_tag}/saltcorn-common.js` },
@@ -156,6 +156,11 @@ const get_headers = (req, version_tag, description, extras = []) => {
       })}`,
     });
   }
+  from_cfg.push({
+    scriptBody: `var dynamic_updates_cfg = ${JSON.stringify({
+      enabled: dynamic_updates_enabled,
+    })}`,
+  });
   return [
     ...stdHeaders,
     ...iconHeader,